requirements.txt
fastapi
uvicorn
python-telegram-bot
pymupdf